balenasign/requirements.txt
Werkzeug==1.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
flask==1.1.2
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.9
pyrsistent==0.17.3
python_gnupg==0.4.6
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
uWSGI==2.0.19.1
Werkzeug==1.0.1
